[MOD] Reworked BaseMemberDrawable into generic base class with value specific subclass BaseMemberValueDrawable<T>
