+++ Assets/GUIUtils/Editor/GUI/Drawables/Members/BaseMemberValueDrawable.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 16556a257bab4beb934ba0cc31b95e95
+timeCreated: 1682063628
